refactor: move cookie locale logic to context
