+++ src/runtime/plugins/i18n.ts
@@ -1,9 +1,8 @@
 import { computed, ref, watch } from 'vue'
 import { createI18n } from 'vue-i18n'
-import { defineNuxtPlugin, prerenderRoutes, useNuxtApp, useRequestHeader } from '#imports'
+import { defineNuxtPlugin, prerenderRoutes, useNuxtApp } from '#imports'
 import { localeCodes, normalizedLocales } from '#build/i18n.options.mjs'
 import { loadAndSetLocale, detectRedirect, navigate, createNuxtI18nDev, createComposableContext } from '../utils'
-import { createI18nCookie } from '../internal'
 import { extendI18n } from '../routing/i18n'
 import { createLogger } from '#nuxt-i18n/logger'
 import { getI18nTarget } from '../compatibility'
@@ -12,7 +11,6 @@ import { useLocalePath, useLocaleRoute, useRouteBaseName, useSwitchLocalePath } 
 import { getDefaultLocaleForDomain, setupMultiDomainLocales } from '../domain'
 import { createLocaleConfigs } from '../shared/locales'
 import { setupVueI18nOptions } from '../shared/vue-i18n'
-import { findBrowserLocale, parseAcceptLanguage } from '#i18n-kit/browser'
 import { createNuxtI18nContext, useLocaleConfigs, useNuxtI18nContext, type NuxtI18nContext } from '../context'
 
 import type { Locale, I18nOptions, Composer, TranslateOptions } from 'vue-i18n'
@@ -73,7 +71,6 @@ export default defineNuxtPlugin({
       nuxt._nuxtI18nDev = createNuxtI18nDev()
     }
 
-    const localeCookie = createI18nCookie()
     // extend i18n instance
     extendI18n(i18n, {
       extendComposer(composer) {
@@ -116,28 +113,9 @@ export default defineNuxtPlugin({
         composer.differentDomains = __DIFFERENT_DOMAINS__
         composer.defaultLocale = runtimeI18n.defaultLocale
 
-        composer.getBrowserLocale = () => {
-          // from navigator or request header
-          const languages = import.meta.client
-            ? navigator.languages
-            : parseAcceptLanguage(useRequestHeader('accept-language'))
-
-          // normalize matching locales
-          const availableLocales = normalizedLocales.map(x => ({ code: x.code, language: x.language ?? x.code }))
-          return findBrowserLocale(availableLocales, languages) || undefined
-        }
-
-        const useCookie = runtimeI18n.detectBrowserLanguage && runtimeI18n.detectBrowserLanguage.useCookie
-        composer.getLocaleCookie = () => {
-          if (useCookie && localeCodes.includes(localeCookie.value || '')) {
-            return localeCookie.value
-          }
-        }
-        composer.setLocaleCookie = (locale: string) => {
-          if (useCookie && localeCodes.includes(locale)) {
-            localeCookie.value = locale
-          }
-        }
+        composer.getBrowserLocale = ctx.getBrowserLocale
+        composer.getLocaleCookie = ctx.getLocaleCookie
+        composer.setLocaleCookie = ctx.setLocaleCookie
 
         composer.onBeforeLanguageSwitch = (oldLocale, newLocale, initialSetup, context) =>
           nuxt.callHook('i18n:beforeLocaleSwitch', {
